requirements.txt
Django==4.2.7
Pillow==10.1.0
PyMySQL==1.1.1
requests==2.31.0
cryptography>=41.0.0
python-dotenv
numpy
geopy
pyproj
python-dateutil
joblib
xgboost
scikit-learn
